+++ geo/src/algorithm/map_coords.rs
@@ -12,14 +12,14 @@
 //! // The NAD83 / California zone 6 (ftUS) coordinate system
 //! let to = "EPSG:2230";
 //! let to_feet = Proj::new_known_crs(&from, &to, None).unwrap();
-//! let f = |x: f64, y: f64| -> Result<_, ProjError> {
+//! let transform = |c: Coordinate<f64>| -> Result<_, ProjError> {
 //!     // proj can accept Point, Coordinate, Tuple, and array values, returning a Result
-//!     let shifted = to_feet.convert((x, y))?;
-//!     Ok((shifted.x(), shifted.y()))
+//!     let shifted = to_feet.convert(c)?;
+//!     Ok(shifted)
 //! };
 //! // 👽
 //! let usa_m = Point::new(-115.797615, 37.2647978);
-//! let usa_ft = usa_m.try_map_coords(|(x, y)| f(x, y)).unwrap();
+//! let usa_ft = usa_m.try_map_coords(|coord| transform(coord)).unwrap();
 //! assert_relative_eq!(6693625.67217475, usa_ft.x(), epsilon = 1e-6);
 //! assert_relative_eq!(3497301.5918027186, usa_ft.y(), epsilon = 1e-6);
 //! ```
@@ -27,7 +27,7 @@
 pub use modern::*;
 mod modern {
     pub(crate) use crate::{
-        coord, CoordNum, Geometry, GeometryCollection, Line, LineString, MultiLineString,
+        CoordNum, Coordinate, Geometry, GeometryCollection, Line, LineString, MultiLineString,
         MultiPoint, MultiPolygon, Point, Polygon, Rect, Triangle,
     };
 
@@ -41,28 +41,39 @@ mod modern {
         ///
         /// ```
         /// use geo::MapCoords;
-        /// use geo::Point;
+        /// use geo::{Coordinate, Point};
         /// use approx::assert_relative_eq;
         ///
         /// let p1 = Point::new(10., 20.);
-        /// let p2 = p1.map_coords(|(x, y)| (x + 1000., y * 2.));
+        /// let p2 = p1.map_coords(|Coordinate { x, y }| Coordinate { x: x + 1000., y: y * 2. });
         ///
         /// assert_relative_eq!(p2, Point::new(1010., 40.), epsilon = 1e-6);
         /// ```
         ///
-        /// You can convert the coordinate type this way as well
+        /// Note that the input and output numeric types need not match.
+        ///
+        /// For example, consider OpenStreetMap's coordinate encoding scheme, which, to save space,
+        /// encodes latitude/longitude as 32bit signed integers from the floating point values
+        /// to six decimal places (eg. lat/lon * 1000000).
         ///
         /// ```
-        /// # use geo::Point;
+        /// # use geo::{Coordinate, Point};
         /// # use geo::MapCoords;
         /// # use approx::assert_relative_eq;
         ///
-        /// let p1: Point<f32> = Point::new(10.0f32, 20.0f32);
-        /// let p2: Point<f64> = p1.map_coords(|(x, y)| (x as f64, y as f64));
+        /// let SCALE_FACTOR: f64 = 1000000.0;
         ///
-        /// assert_relative_eq!(p2, Point::new(10.0f64, 20.0f64), epsilon = 1e-6);
+        /// let floating_point_geom: Point<f64> = Point::new(10.15f64, 20.05f64);
+        /// let fixed_point_geom: Point<i32> = floating_point_geom.map_coords(|Coordinate { x, y }| {
+        ///     Coordinate { x: (x * SCALE_FACTOR) as i32, y: (y * SCALE_FACTOR) as i32 }
+        /// });
+        ///
+        /// assert_eq!(fixed_point_geom.x(), 10150000);
         /// ```
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output
+        ///
+        /// If you want *only* to convert between numeric types (i32 -> f64) without further
+        /// transformation, consider using [`Convert].
+        fn map_coords(&self, func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy) -> Self::Output
         where
             T: CoordNum,
             NT: CoordNum;
@@ -74,12 +85,12 @@ mod modern {
         /// ```
         /// use approx::assert_relative_eq;
         /// use geo::MapCoords;
-        /// use geo::Point;
+        /// use geo::{Coordinate, Point};
         ///
         /// let p1 = Point::new(10., 20.);
         /// let p2 = p1
-        ///     .try_map_coords(|(x, y)| -> Result<_, std::convert::Infallible> {
-        ///         Ok((x + 1000., y * 2.))
+        ///     .try_map_coords(|Coordinate { x, y }| -> Result<_, std::convert::Infallible> {
+        ///         Ok(Coordinate { x: x + 1000., y: y * 2. })
         ///     }).unwrap();
         ///
         /// assert_relative_eq!(p2, Point::new(1010., 40.), epsilon = 1e-6);
@@ -99,20 +110,20 @@ mod modern {
         /// // The NAD83 / California zone 6 (ftUS) coordinate system
         /// let to = "EPSG:2230";
         /// let to_feet = Proj::new_known_crs(&from, &to, None).unwrap();
-        /// let f = |x: f64, y: f64| -> Result<_, ProjError> {
+        /// let transform = |c: Coordinate<f64>| -> Result<_, ProjError> {
         ///     // proj can accept Point, Coordinate, Tuple, and array values, returning a Result
-        ///     let shifted = to_feet.convert((x, y))?;
-        ///     Ok((shifted.x(), shifted.y()))
+        ///     let shifted = to_feet.convert(c)?;
+        ///     Ok(shifted)
         /// };
         /// // 👽
         /// let usa_m = Point::new(-115.797615, 37.2647978);
-        /// let usa_ft = usa_m.try_map_coords(|(x, y)| f(x, y)).unwrap();
+        /// let usa_ft = usa_m.try_map_coords(|coord| transform(coord)).unwrap();
         /// assert_relative_eq!(6693625.67217475, usa_ft.x(), epsilon = 1e-6);
         /// assert_relative_eq!(3497301.5918027186, usa_ft.y(), epsilon = 1e-6);
         /// ```
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E>
         where
             T: CoordNum,
@@ -126,15 +137,15 @@ mod modern {
         ///
         /// ```
         /// use geo::MapCoordsInPlace;
-        /// use geo::Point;
+        /// use geo::{Coordinate, Point};
         /// use approx::assert_relative_eq;
         ///
         /// let mut p = Point::new(10., 20.);
-        /// p.map_coords_in_place(|(x, y)| (x + 1000., y * 2.));
+        /// p.map_coords_in_place(|Coordinate { x, y }| Coordinate { x: x + 1000., y: y * 2. });
         ///
         /// assert_relative_eq!(p, Point::new(1010., 40.), epsilon = 1e-6);
         /// ```
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T) + Copy)
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T> + Copy)
         where
             T: CoordNum;
 
@@ -147,14 +158,15 @@ mod modern {
         ///
         /// ```
         /// use geo::MapCoordsInPlace;
+        /// use geo::Coordinate;
         ///
         /// let mut p1 = geo::point!{x: 10u32, y: 20u32};
         ///
-        /// p1.try_map_coords_in_place(|(x, y)| -> Result<_, &str> {
-        ///     Ok((
-        ///         x.checked_add(1000).ok_or("Overflow")?,
-        ///         y.checked_mul(2).ok_or("Overflow")?,
-        ///     ))
+        /// p1.try_map_coords_in_place(|Coordinate { x, y }| -> Result<_, &str> {
+        ///     Ok(Coordinate {
+        ///         x: x.checked_add(1000).ok_or("Overflow")?,
+        ///         y: y.checked_mul(2).ok_or("Overflow")?,
+        ///     })
         /// })?;
         ///
         /// assert_eq!(
@@ -165,7 +177,7 @@ mod modern {
         /// ```
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E>
         where
             T: CoordNum;
@@ -178,35 +190,31 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for Point<T> {
         type Output = Point<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
-            let new_point = func((self.0.x, self.0.y));
-            Point::new(new_point.0, new_point.1)
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
+            Point(func(self.0))
         }
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E>,
         ) -> Result<Self::Output, E> {
-            let new_point = func((self.0.x, self.0.y))?;
-            Ok(Point::new(new_point.0, new_point.1))
+            Ok(Point(func(self.0)?))
         }
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for Point<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T)) {
-            let new_point = func((self.0.x, self.0.y));
-            self.0.x = new_point.0;
-            self.0.y = new_point.1;
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T>) {
+            self.0 = func(self.0);
         }
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
-            let new_point = func((self.0.x, self.0.y))?;
-            self.0.x = new_point.0;
-            self.0.y = new_point.1;
-
+            self.0 = func(self.0)?;
             Ok(())
         }
     }
@@ -218,7 +226,10 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for Line<T> {
         type Output = Line<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
             Line::new(
                 self.start_point().map_coords(func).0,
                 self.end_point().map_coords(func).0,
@@ -227,7 +238,7 @@ mod modern {
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E> {
             Ok(Line::new(
                 self.start_point().try_map_coords(func)?.0,
@@ -237,27 +248,17 @@ mod modern {
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for Line<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T)) {
-            let new_start = func((self.start.x, self.start.y));
-            self.start.x = new_start.0;
-            self.start.y = new_start.1;
-
-            let new_end = func((self.end.x, self.end.y));
-            self.end.x = new_end.0;
-            self.end.y = new_end.1;
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T>) {
+            self.start = func(self.start);
+            self.end = func(self.end);
         }
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
-            let new_start = func((self.start.x, self.start.y))?;
-            self.start.x = new_start.0;
-            self.start.y = new_start.1;
-
-            let new_end = func((self.end.x, self.end.y))?;
-            self.end.x = new_end.0;
-            self.end.y = new_end.1;
+            self.start = func(self.start)?;
+            self.end = func(self.end)?;
 
             Ok(())
         }
@@ -270,7 +271,10 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for LineString<T> {
         type Output = LineString<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
             LineString::from(
                 self.points()
                     .map(|p| p.map_coords(func))
@@ -280,7 +284,7 @@ mod modern {
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E> {
             Ok(LineString::from(
                 self.points()
@@ -291,22 +295,18 @@ mod modern {
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for LineString<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T)) {
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T>) {
             for p in &mut self.0 {
-                let new_coords = func((p.x, p.y));
-                p.x = new_coords.0;
-                p.y = new_coords.1;
+                *p = func(*p);
             }
         }
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
             for p in &mut self.0 {
-                let new_coords = func((p.x, p.y))?;
-                p.x = new_coords.0;
-                p.y = new_coords.1;
+                *p = func(*p)?;
             }
             Ok(())
         }
@@ -319,7 +319,10 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for Polygon<T> {
         type Output = Polygon<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
             Polygon::new(
                 self.exterior().map_coords(func),
                 self.interiors()
@@ -331,7 +334,7 @@ mod modern {
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E> {
             Ok(Polygon::new(
                 self.exterior().try_map_coords(func)?,
@@ -344,7 +347,7 @@ mod modern {
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for Polygon<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T) + Copy) {
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T> + Copy) {
             self.exterior_mut(|line_string| {
                 line_string.map_coords_in_place(func);
             });
@@ -358,7 +361,7 @@ mod modern {
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
             let mut result = Ok(());
 
@@ -390,13 +393,16 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for MultiPoint<T> {
         type Output = MultiPoint<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
             MultiPoint::new(self.iter().map(|p| p.map_coords(func)).collect())
         }
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E> {
             Ok(MultiPoint::new(
                 self.0
@@ -408,7 +414,7 @@ mod modern {
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for MultiPoint<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T) + Copy) {
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T> + Copy) {
             for p in &mut self.0 {
                 p.map_coords_in_place(func);
             }
@@ -416,7 +422,7 @@ mod modern {
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
             for p in &mut self.0 {
                 p.try_map_coords_in_place(&func)?;
@@ -432,13 +438,16 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for MultiLineString<T> {
         type Output = MultiLineString<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
             MultiLineString::new(self.iter().map(|l| l.map_coords(func)).collect())
         }
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E> {
             Ok(MultiLineString::new(
                 self.0
@@ -450,7 +459,7 @@ mod modern {
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for MultiLineString<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T) + Copy) {
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T> + Copy) {
             for p in &mut self.0 {
                 p.map_coords_in_place(func);
             }
@@ -458,7 +467,7 @@ mod modern {
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
             for p in &mut self.0 {
                 p.try_map_coords_in_place(&func)?;
@@ -474,13 +483,16 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for MultiPolygon<T> {
         type Output = MultiPolygon<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
             MultiPolygon::new(self.iter().map(|p| p.map_coords(func)).collect())
         }
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E> {
             Ok(MultiPolygon::new(
                 self.0
@@ -492,7 +504,7 @@ mod modern {
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for MultiPolygon<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T) + Copy) {
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T> + Copy) {
             for p in &mut self.0 {
                 p.map_coords_in_place(func);
             }
@@ -500,7 +512,7 @@ mod modern {
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
             for p in &mut self.0 {
                 p.try_map_coords_in_place(&func)?;
@@ -516,7 +528,10 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for Geometry<T> {
         type Output = Geometry<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
             match *self {
                 Geometry::Point(ref x) => Geometry::Point(x.map_coords(func)),
                 Geometry::Line(ref x) => Geometry::Line(x.map_coords(func)),
@@ -535,7 +550,7 @@ mod modern {
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E> {
             match *self {
                 Geometry::Point(ref x) => Ok(Geometry::Point(x.try_map_coords(func)?)),
@@ -559,7 +574,7 @@ mod modern {
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for Geometry<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T) + Copy) {
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T> + Copy) {
             match *self {
                 Geometry::Point(ref mut x) => x.map_coords_in_place(func),
                 Geometry::Line(ref mut x) => x.map_coords_in_place(func),
@@ -576,7 +591,7 @@ mod modern {
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
             match *self {
                 Geometry::Point(ref mut x) => x.try_map_coords_in_place(func),
@@ -600,13 +615,16 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for GeometryCollection<T> {
         type Output = GeometryCollection<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
             GeometryCollection::new_from(self.iter().map(|g| g.map_coords(func)).collect())
         }
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E> + Copy,
         ) -> Result<Self::Output, E> {
             Ok(GeometryCollection::new_from(
                 self.0
@@ -618,7 +636,7 @@ mod modern {
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for GeometryCollection<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T) + Copy) {
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T> + Copy) {
             for p in &mut self.0 {
                 p.map_coords_in_place(func);
             }
@@ -626,7 +644,7 @@ mod modern {
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
             for p in &mut self.0 {
                 p.try_map_coords_in_place(&func)?;
@@ -642,29 +660,32 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for Rect<T> {
         type Output = Rect<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
-            Rect::new(func(self.min().x_y()), func(self.max().x_y()))
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
+            Rect::new(func(self.min()), func(self.max()))
         }
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E>,
         ) -> Result<Self::Output, E> {
-            Ok(Rect::new(func(self.min().x_y())?, func(self.max().x_y())?))
+            Ok(Rect::new(func(self.min())?, func(self.max())?))
         }
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for Rect<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T)) {
-            let mut new_rect = Rect::new(func(self.min().x_y()), func(self.max().x_y()));
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T>) {
+            let mut new_rect = Rect::new(func(self.min()), func(self.max()));
             ::std::mem::swap(self, &mut new_rect);
         }
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
-            let mut new_rect = Rect::new(func(self.min().x_y())?, func(self.max().x_y())?);
+            let mut new_rect = Rect::new(func(self.min())?, func(self.max())?);
             ::std::mem::swap(self, &mut new_rect);
             Ok(())
         }
@@ -677,62 +698,33 @@ mod modern {
     impl<T: CoordNum, NT: CoordNum> MapCoords<T, NT> for Triangle<T> {
         type Output = Triangle<NT>;
 
-        fn map_coords(&self, func: impl Fn((T, T)) -> (NT, NT) + Copy) -> Self::Output {
-            let p1 = func(self.0.x_y());
-            let p2 = func(self.1.x_y());
-            let p3 = func(self.2.x_y());
-
-            Triangle::new(
-                coord! { x: p1.0, y: p1.1 },
-                coord! { x: p2.0, y: p2.1 },
-                coord! { x: p3.0, y: p3.1 },
-            )
+        fn map_coords(
+            &self,
+            func: impl Fn(Coordinate<T>) -> Coordinate<NT> + Copy,
+        ) -> Self::Output {
+            Triangle::new(func(self.0), func(self.1), func(self.2))
         }
 
         fn try_map_coords<E>(
             &self,
-            func: impl Fn((T, T)) -> Result<(NT, NT), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<NT>, E>,
         ) -> Result<Self::Output, E> {
-            let p1 = func(self.0.x_y())?;
-            let p2 = func(self.1.x_y())?;
-            let p3 = func(self.2.x_y())?;
-
-            Ok(Triangle::new(
-                coord! { x: p1.0, y: p1.1 },
-                coord! { x: p2.0, y: p2.1 },
-                coord! { x: p3.0, y: p3.1 },
-            ))
+            Ok(Triangle::new(func(self.0)?, func(self.1)?, func(self.2)?))
         }
     }
 
     impl<T: CoordNum> MapCoordsInPlace<T> for Triangle<T> {
-        fn map_coords_in_place(&mut self, func: impl Fn((T, T)) -> (T, T)) {
-            let p1 = func(self.0.x_y());
-            let p2 = func(self.1.x_y());
-            let p3 = func(self.2.x_y());
-
-            let mut new_triangle = Triangle::new(
-                coord! { x: p1.0, y: p1.1 },
-                coord! { x: p2.0, y: p2.1 },
-                coord! { x: p3.0, y: p3.1 },
-            );
+        fn map_coords_in_place(&mut self, func: impl Fn(Coordinate<T>) -> Coordinate<T>) {
+            let mut new_triangle = Triangle::new(func(self.0), func(self.1), func(self.2));
 
             ::std::mem::swap(self, &mut new_triangle);
         }
 
         fn try_map_coords_in_place<E>(
             &mut self,
-            func: impl Fn((T, T)) -> Result<(T, T), E>,
+            func: impl Fn(Coordinate<T>) -> Result<Coordinate<T>, E>,
         ) -> Result<(), E> {
-            let p1 = func(self.0.x_y())?;
-            let p2 = func(self.1.x_y())?;
-            let p3 = func(self.2.x_y())?;
-
-            let mut new_triangle = Triangle::new(
-                coord! { x: p1.0, y: p1.1 },
-                coord! { x: p2.0, y: p2.1 },
-                coord! { x: p3.0, y: p3.1 },
-            );
+            let mut new_triangle = Triangle::new(func(self.0)?, func(self.1)?, func(self.2)?);
 
             ::std::mem::swap(self, &mut new_triangle);
 
@@ -745,7 +737,10 @@ pub(crate) mod deprecated {
     use super::*;
 
     /// Map a fallible function over all the coordinates in a geometry, returning a Result
-    #[deprecated(since = "0.21.0", note = "use `MapCoords::try_map_coords` instead")]
+    #[deprecated(
+        since = "0.21.0",
+        note = "use `MapCoords::try_map_coords` which takes a `Coordinate` instead of an (x,y) tuple"
+    )]
     pub trait TryMapCoords<T, NT, E> {
         type Output;
 
@@ -808,7 +803,7 @@ pub(crate) mod deprecated {
 
     #[deprecated(
         since = "0.21.0",
-        note = "use `MapCoordsInPlace::try_map_coords_in_place` instead"
+        note = "use `MapCoordsInPlace::try_map_coords_in_place` which takes a `Coordinate` instead of an (x,y) tuple"
     )]
     pub trait TryMapCoordsInplace<T, E> {
         /// Map a fallible function over all the coordinates in a geometry, in place, returning a `Result`.
@@ -849,7 +844,7 @@ pub(crate) mod deprecated {
     /// Map a function over all the coordinates in an object in place
     #[deprecated(
         since = "0.21.0",
-        note = "use `MapCoordsInPlace::map_coords_in_place` instead"
+        note = "use `MapCoordsInPlace::map_coords_in_place` instead which takes a `Coordinate` instead of an (x,y) tuple"
     )]
     pub trait MapCoordsInplace<T>: MapCoordsInPlace<T> {
         /// Apply a function to all the coordinates in a geometric object, in place
@@ -883,7 +878,7 @@ pub(crate) mod deprecated {
                     &self,
                     func: impl Fn((T, T)) -> Result<(NT, NT), E> + Copy,
                 ) -> Result<Self::Output, E> {
-                    MapCoords::try_map_coords(self, func)
+                    MapCoords::try_map_coords(self, |c| Ok(Coordinate::from(func(c.x_y())?)))
                 }
             }
 
@@ -893,7 +888,7 @@ pub(crate) mod deprecated {
                     &mut self,
                     func: impl Fn((T, T)) -> Result<(T, T), E>,
                 ) -> Result<(), E> {
-                    MapCoordsInPlace::try_map_coords_in_place(self, func)
+                    MapCoordsInPlace::try_map_coords_in_place(self, |c| Ok(func(c.x_y())?.into()))
                 }
             }
 
@@ -919,7 +914,7 @@ pub(crate) mod deprecated {
                 where
                     T: CoordNum,
                 {
-                    MapCoordsInPlace::map_coords_in_place(self, func)
+                    MapCoordsInPlace::map_coords_in_place(self, |c| func(c.x_y()).into())
                 }
             }
         };
@@ -949,7 +944,7 @@ mod test {
     #[test]
     fn point() {
         let p = Point::new(10., 10.);
-        let new_p = p.map_coords(|(x, y)| (x + 10., y + 100.));
+        let new_p = p.map_coords(|Coordinate { x, y }| (x + 10., y + 100.).into());
         assert_relative_eq!(new_p.x(), 20.);
         assert_relative_eq!(new_p.y(), 110.);
     }
@@ -957,7 +952,7 @@ mod test {
     #[test]
     fn point_inplace() {
         let mut p2 = Point::new(10f32, 10f32);
-        p2.map_coords_in_place(|(x, y)| (x + 10., y + 100.));
+        p2.map_coords_in_place(|Coordinate { x, y }| (x + 10., y + 100.).into());
         assert_relative_eq!(p2.x(), 20.);
         assert_relative_eq!(p2.y(), 110.);
     }
@@ -965,7 +960,7 @@ mod test {
     #[test]
     fn rect_inplace() {
         let mut rect = Rect::new((10, 10), (20, 20));
-        rect.map_coords_in_place(|(x, y)| (x + 10, y + 20));
+        rect.map_coords_in_place(|Coordinate { x, y }| (x + 10, y + 20).into());
         assert_eq!(rect.min(), coord! { x: 20, y: 30 });
         assert_eq!(rect.max(), coord! { x: 30, y: 40 });
     }
@@ -976,11 +971,11 @@ mod test {
         // Rect's enforce that rect.min is up and left of p2.  Here we test that the points are
         // normalized into a valid rect, regardless of the order they are mapped.
         rect.map_coords_in_place(|pt| {
-            match pt {
+            match pt.x_y() {
                 // old min point maps to new max point
-                (2, 2) => (4, 4),
+                (2, 2) => (4, 4).into(),
                 // old max point maps to new min point
-                (3, 3) => (1, 1),
+                (3, 3) => (1, 1).into(),
                 _ => panic!("unexpected point"),
             }
         });
@@ -992,7 +987,7 @@ mod test {
     #[test]
     fn rect_map_coords() {
         let rect = Rect::new((10, 10), (20, 20));
-        let another_rect = rect.map_coords(|(x, y)| (x + 10, y + 20));
+        let another_rect = rect.map_coords(|Coordinate { x, y }| (x + 10, y + 20).into());
         assert_eq!(another_rect.min(), coord! { x: 20, y: 30 });
         assert_eq!(another_rect.max(), coord! { x: 30, y: 40 });
     }
@@ -1000,11 +995,12 @@ mod test {
     #[test]
     fn rect_try_map_coords() {
         let rect = Rect::new((10i32, 10), (20, 20));
-        let result = rect.try_map_coords(|(x, y)| -> Result<_, &'static str> {
+        let result = rect.try_map_coords(|Coordinate { x, y }| -> Result<_, &'static str> {
             Ok((
                 x.checked_add(10).ok_or("overflow")?,
                 y.checked_add(20).ok_or("overflow")?,
-            ))
+            )
+                .into())
         });
         assert!(result.is_ok());
     }
@@ -1015,11 +1011,11 @@ mod test {
         // Rect's enforce that rect.min is up and left of p2.  Here we test that the points are
         // normalized into a valid rect, regardless of the order they are mapped.
         let result: Result<_, std::convert::Infallible> = rect.try_map_coords(|pt| {
-            match pt {
+            match pt.x_y() {
                 // old min point maps to new max point
-                (2, 2) => Ok((4, 4)),
+                (2, 2) => Ok((4, 4).into()),
                 // old max point maps to new min point
-                (3, 3) => Ok((1, 1)),
+                (3, 3) => Ok((1, 1).into()),
                 _ => panic!("unexpected point"),
             }
         });
@@ -1032,7 +1028,7 @@ mod test {
     fn line() {
         let line = Line::from([(0., 0.), (1., 2.)]);
         assert_relative_eq!(
-            line.map_coords(|(x, y)| (x * 2., y)),
+            line.map_coords(|Coordinate { x, y }| (x * 2., y).into()),
             Line::from([(0., 0.), (2., 2.)]),
             epsilon = 1e-6
         );
@@ -1041,7 +1037,7 @@ mod test {
     #[test]
     fn linestring() {
         let line1: LineString<f32> = LineString::from(vec![(0., 0.), (1., 2.)]);
-        let line2 = line1.map_coords(|(x, y)| (x + 10., y - 100.));
+        let line2 = line1.map_coords(|Coordinate { x, y }| (x + 10., y - 100.).into());
         assert_relative_eq!(line2.0[0], Coordinate::from((10., -100.)), epsilon = 1e-6);
         assert_relative_eq!(line2.0[1], Coordinate::from((11., -98.)), epsilon = 1e-6);
     }
@@ -1057,7 +1053,7 @@ mod test {
         ])];
         let p = Polygon::new(exterior, interiors);
 
-        let p2 = p.map_coords(|(x, y)| (x + 10., y - 100.));
+        let p2 = p.map_coords(|Coordinate { x, y }| (x + 10., y - 100.).into());
 
         let exterior2 =
             LineString::from(vec![(10., -100.), (11., -99.), (11., -100.), (10., -100.)]);
@@ -1079,7 +1075,7 @@ mod test {
         let mp = MultiPoint::new(vec![p1, p2]);
 
         assert_eq!(
-            mp.map_coords(|(x, y)| (x + 10., y + 100.)),
+            mp.map_coords(|Coordinate { x, y }| (x + 10., y + 100.).into()),
             MultiPoint::new(vec![Point::new(20., 110.), Point::new(10., 0.)])
         );
     }
@@ -1089,7 +1085,7 @@ mod test {
         let line1: LineString<f32> = LineString::from(vec![(0., 0.), (1., 2.)]);
         let line2: LineString<f32> = LineString::from(vec![(-1., 0.), (0., 0.), (1., 2.)]);
         let mline = MultiLineString::new(vec![line1, line2]);
-        let mline2 = mline.map_coords(|(x, y)| (x + 10., y - 100.));
+        let mline2 = mline.map_coords(|Coordinate { x, y }| (x + 10., y - 100.).into());
         assert_relative_eq!(
             mline2,
             MultiLineString::new(vec![
@@ -1129,7 +1125,7 @@ mod test {
         ];
 
         let mp = MultiPolygon::new(vec![poly1, poly2]);
-        let mp2 = mp.map_coords(|(x, y)| (x * 2., y + 100.));
+        let mp2 = mp.map_coords(|Coordinate { x, y }| (x * 2., y + 100.).into());
         assert_eq!(mp2.0.len(), 2);
         assert_relative_eq!(
             mp2.0[0],
@@ -1172,7 +1168,7 @@ mod test {
         let gc = GeometryCollection::new_from(vec![p1, line1]);
 
         assert_eq!(
-            gc.map_coords(|(x, y)| (x + 10., y + 100.)),
+            gc.map_coords(|Coordinate { x, y }| (x + 10., y + 100.).into()),
             GeometryCollection::new_from(vec![
                 Geometry::Point(Point::new(20., 110.)),
                 Geometry::LineString(LineString::from(vec![(10., 100.), (11., 102.)])),
@@ -1183,7 +1179,7 @@ mod test {
     #[test]
     fn convert_type() {
         let p1: Point<f64> = Point::new(1., 2.);
-        let p2: Point<f32> = p1.map_coords(|(x, y)| (x as f32, y as f32));
+        let p2: Point<f32> = p1.map_coords(|Coordinate { x, y }| (x as f32, y as f32).into());
         assert_relative_eq!(p2.x(), 1f32);
         assert_relative_eq!(p2.y(), 2f32);
     }
@@ -1191,27 +1187,27 @@ mod test {
     #[cfg(feature = "use-proj")]
     #[test]
     fn test_fallible_proj() {
-        use proj::{Coord, Proj, ProjError};
+        use proj::{Proj, ProjError};
         let from = "EPSG:4326";
         let to = "EPSG:2230";
         let to_feet = Proj::new_known_crs(from, to, None).unwrap();
 
-        let f = |x: f64, y: f64| -> Result<_, ProjError> {
-            let shifted = to_feet.convert((x, y))?;
-            Ok((shifted.x(), shifted.y()))
+        let f = |c| -> Result<_, ProjError> {
+            let shifted = to_feet.convert(c)?;
+            Ok(shifted)
         };
         // 👽
         let usa_m = Point::new(-115.797615, 37.2647978);
-        let usa_ft = usa_m.try_map_coords(|(x, y)| f(x, y)).unwrap();
+        let usa_ft = usa_m.try_map_coords(f).unwrap();
         assert_relative_eq!(6693625.67217475, usa_ft.x(), epsilon = 1e-6);
         assert_relative_eq!(3497301.5918027186, usa_ft.y(), epsilon = 1e-6);
     }
 
     #[test]
     fn test_fallible() {
-        let f = |x: f64, y: f64| -> Result<_, &'static str> {
+        let f = |Coordinate { x, y }| -> Result<_, &'static str> {
             if relative_ne!(x, 2.0) {
-                Ok((x * 2., y + 100.))
+                Ok((x * 2., y + 100.).into())
             } else {
                 Err("Ugh")
             }
@@ -1230,9 +1226,9 @@ mod test {
             Point::new(3.0, 3.0),
         ]
         .into();
-        let bad = bad_ls.try_map_coords(|(x, y)| f(x, y));
+        let bad = bad_ls.try_map_coords(f);
         assert!(bad.is_err());
-        let good = good_ls.try_map_coords(|(x, y)| f(x, y));
+        let good = good_ls.try_map_coords(f);
         assert!(good.is_ok());
         assert_relative_eq!(
             good.unwrap(),
@@ -1251,6 +1247,6 @@ mod test {
 
         // This call should not panic even though Rect::new
         // constructor panics if min coords > max coords
-        rect.map_coords(|(x, y)| (-x, -y));
+        rect.map_coords(|Coordinate { x, y }| (-x, -y).into());
     }
 }